+++ CHANGELOG.md
@@ -11,7 +11,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [0.2.0] - 
 ### Added
-- `title` and `titleWidth` parameters for `dashboard_header`
+- `title`, `titleWidth` and `menu_button` parameters for `dashboardHeader`
+- `overlay` and `dim_page` parameters for `dashboardSidebar`
+- `fluid` param for `tabItem`
+
+### Changed
+- Main content will shrink, not move when sidebar is opened and `overlay = FALSE`
+### Fixed
+- Issues with layout responsiveness
+- Issues with sidebar position and overflow
 
 ## [0.1.5] - 2020-03-30
 ### Added